requirements.txt
requests>=2.32.0
python-dotenv>=1.0.1

pillow>=10.4.0
fpdf2>=2.7.9

# video + upload
moviepy>=1.0.3
edge-tts>=6.1.12
google-api-python-client>=2.130.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
